requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
zipp==3.20.2
